Add shader for shape
* shape properties with border radius

* fix boder radius

* fix depth on ui

+++ src/Sea/Core/VideoMode.hpp
@@ -12,8 +12,8 @@ namespace Sea
 
     struct VideoMode
     {
-		u32 Width = 1080;
-		u32 Height = 720;
+		f32 Width = 1080;
+		f32 Height = 720;
 		bool Resizable = false;
 		bool Fullscreen = false;
 		bool Maximazed = false;
